+++ packages/reactive-element/src/css-tag.ts
@@ -15,21 +15,36 @@ export const supportsAdoptingStyleSheets =
   'adoptedStyleSheets' in Document.prototype &&
   'replace' in CSSStyleSheet.prototype;
 
+/**
+ * A CSSResult or native CSSStyleSheet.
+ *
+ * In browsers that support constructible CSS style sheets, CSSStyleSheet
+ * object can be used for styling along side CSSResult from the `css`
+ * template tag.
+ */
 export type CSSResultOrNative = CSSResult | CSSStyleSheet;
-
-export type CSSResultFlatArray = CSSResultOrNative[];
 
 export type CSSResultArray = Array<CSSResultOrNative | CSSResultArray>;
 
+/**
+ * A single CSSResult, CSSStyleSheet, or an array or nested arrays of those.
+ */
 export type CSSResultGroup = CSSResultOrNative | CSSResultArray;
 
 const constructionToken = Symbol();
 
+/**
+ * A container for a string of CSS text, that may be used to create a CSSStyleSheet.
+ *
+ * CSSResult is the return value of `css`-tagged template literals and
+ * `unsafeCSS()`. In order to ensure that CSSResults are only created via the
+ * `css` tag and `unsafeCSS()`, CSSResult cannot be constructed directly.
+ */
 export class CSSResult {
   readonly cssText: string;
   private _styleSheet?: CSSStyleSheet;
 
-  constructor(cssText: string, safeToken: symbol) {
+  private constructor(cssText: string, safeToken: symbol) {
     if (safeToken !== constructionToken) {
       throw new Error(
         'CSSResult is not constructable. Use `unsafeCSS` or `css` instead.'
@@ -55,6 +70,10 @@ export class CSSResult {
   }
 }
 
+type ConstructableCSSResult = CSSResult & {
+  new (cssText: string, safeToken: symbol): CSSResult;
+};
+
 const cssResultCache = new Map<string, CSSResult>();
 
 const getCSSResult = (cssText: string): CSSResult => {
@@ -62,7 +81,10 @@ const getCSSResult = (cssText: string): CSSResult => {
   if (result === undefined) {
     cssResultCache.set(
       cssText,
-      (result = new CSSResult(cssText, constructionToken))
+      (result = new (CSSResult as ConstructableCSSResult)(
+        cssText,
+        constructionToken
+      ))
     );
   }
   return result;
@@ -94,15 +116,17 @@ export const unsafeCSS = (value: unknown) => {
 };
 
 /**
- * Template tag which which can be used with LitElement's [[LitElement.styles |
- * `styles`]] property to set element styles. For security reasons, only literal
- * string values may be used. To incorporate non-literal values [[`unsafeCSS`]]
- * may be used inside a template string part.
+ * A template literal tag which can be used with LitElement's
+ * [[LitElement.styles | `styles`]] property to set element styles.
+ *
+ * For security reasons, only literal string values and number may be used in
+ * embedded expressions. To incorporate non-literal values [[`unsafeCSS`]] may
+ * be used inside an expression.
  */
 export const css = (
   strings: TemplateStringsArray,
   ...values: (CSSResultGroup | number)[]
-): CSSResultGroup => {
+): CSSResult => {
   const cssText =
     strings.length === 1
       ? strings[0]
@@ -124,7 +148,7 @@ export const css = (
  */
 export const adoptStyles = (
   renderRoot: ShadowRoot,
-  styles: CSSResultFlatArray
+  styles: Array<CSSResultOrNative>
 ) => {
   if (supportsAdoptingStyleSheets) {
     (renderRoot as ShadowRoot).adoptedStyleSheets = styles.map((s) =>